requirements.txt
scapy==2.5.0
aiohttp==3.9.3
aiodns==3.1.1
tldextract==5.1.2
jinja2==3.1.3
netifaces==0.11.0
beautifulsoup4==4.12.3
